+++ karma.conf.ts
@@ -3,9 +3,7 @@ import * as path from 'path';
 import * as ForkTsCheckerWebpackPlugin from 'fork-ts-checker-webpack-plugin';
 
 export default (config: any) => {
-
   config.set({
-
     // base path that will be used to resolve all patterns (eg. files, exclude)
     basePath: './',
 
@@ -14,9 +12,7 @@ export default (config: any) => {
     frameworks: ['mocha'],
 
     // list of files / patterns to load in the browser
-    files: [
-      'test/entry.ts'
-    ],
+    files: ['test/entry.ts'],
 
     // preprocess matching files before serving them to the browser
     // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
@@ -29,30 +25,34 @@ export default (config: any) => {
         extensions: ['.ts', '.js']
       },
       module: {
-        rules: [{
-          test: /\.ts$/,
-          loader: 'tslint-loader',
-          exclude: /node_modules/,
-          enforce: 'pre',
-          options: {
-            tslint: {
-              emitErrors: config.singleRun,
-              failOnHint: config.singleRun
+        rules: [
+          {
+            test: /\.ts$/,
+            loader: 'tslint-loader',
+            exclude: /node_modules/,
+            enforce: 'pre',
+            options: {
+              tslint: {
+                emitErrors: config.singleRun,
+                failOnHint: config.singleRun
+              }
             }
+          },
+          {
+            test: /\.ts$/,
+            loader: 'ts-loader',
+            exclude: /node_modules/,
+            options: {
+              transpileOnly: true
+            }
+          },
+          {
+            test: /src\/.+\.ts$/,
+            exclude: /(node_modules|\.spec\.ts$)/,
+            loader: 'istanbul-instrumenter-loader',
+            enforce: 'post'
           }
-        }, {
-          test: /\.ts$/,
-          loader: 'ts-loader',
-          exclude: /node_modules/,
-          options: {
-            transpileOnly: true
-          }
-        }, {
-          test: /src\/.+\.ts$/,
-          exclude: /(node_modules|\.spec\.ts$)/,
-          loader: 'istanbul-instrumenter-loader',
-          enforce: 'post'
-        }]
+        ]
       },
       plugins: [
         new webpack.SourceMapDevToolPlugin({
@@ -98,7 +98,5 @@ export default (config: any) => {
     // start these browsers
     // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
     browsers: ['ChromeHeadless']
-
   });
-
 };