+++ src/flatpickr.directive.ts
@@ -10,7 +10,8 @@ import {
   OnDestroy,
   forwardRef,
   HostListener,
-  Inject
+  Inject,
+  Renderer2
 } from '@angular/core';
 import {
   FlatpickrDefaults,
@@ -38,7 +39,11 @@ export const FLATPICKR_CONTROL_VALUE_ACCESSOR: any = {
 
 @Directive({
   selector: '[mwlFlatpickr]',
-  providers: [FLATPICKR_CONTROL_VALUE_ACCESSOR]
+  providers: [FLATPICKR_CONTROL_VALUE_ACCESSOR],
+  host: {
+    // tslint:disable-line use-host-property-decorator
+    '(blur)': 'onTouchedFn()'
+  }
 })
 export class FlatpickrDirective
   implements AfterViewInit, OnChanges, OnDestroy, ControlValueAccessor {
@@ -264,9 +269,12 @@ export class FlatpickrDirective
 
   onChangeFn: (value: any) => void = () => {}; // tslint:disable-line
 
+  onTouchedFn = () => {};
+
   constructor(
     private elm: ElementRef,
     private defaults: FlatpickrDefaults,
+    private renderer: Renderer2,
     @Inject(FLATPICKR) private flatpickr
   ) {}
 
@@ -392,7 +400,13 @@ export class FlatpickrDirective
     this.onChangeFn = fn;
   }
 
-  registerOnTouched(fn: any): void {} // tslint:disable-line
+  registerOnTouched(fn: () => void): void {
+    this.onTouchedFn = fn;
+  }
+
+  setDisabledState(isDisabled: boolean): void {
+    this.renderer.setProperty(this.elm.nativeElement, 'disabled', isDisabled);
+  }
 
   @HostListener('input')
   inputChanged(): void {